Move G2 utils to the `ui` folder
