+++ packages/components/src/ui/text/use-text.js
@@ -14,7 +14,7 @@ import { useMemo, Children, cloneElement } from '@wordpress/element';
  * Internal dependencies
  */
 import { useTruncate } from '../truncate';
-import { getOptimalTextShade } from '../../utils/colors';
+import { getOptimalTextShade } from '../utils';
 import * as styles from './styles';
 import { createHighlighterText } from './utils';
