+++ packages/components/src/ui/utils/index.js
@@ -1,1 +1,2 @@
+export { getOptimalTextShade } from './colors';
 export { createComponent } from './create-component';
